Reorganize mudd/services/ directory to separate concerns
Move non-service modules out of services/ package to improve code organization:

- Create mudd/types.py for VerbAction enum
- Create mudd/database.py combining database pool and migrations
- Create mudd/loaders/ package for data sync functions (zone_loader, entity_loader, verb_loader)
- Create mudd/matching/ package for matching functions (entity_matcher, verb_matcher)
- Merge VisibilityServiceProtocol into visibility.py
- Add CLAUDE.md documentation to services/, loaders/, and matching/ packages

The services/ package now only contains DI service classes that manage runtime state and caching.
