+++ justfile
@@ -54,4 +54,4 @@ resetdb:
     psql postgresql://mudd:mudd@db:5432/mudd -c "DROP SCHEMA public CASCADE; CREATE SCHEMA public;"
 
 migratedb:
-    uv run python -c "import asyncio; from mudd.services.database import init_database; asyncio.run(init_database())"
+    uv run python -c "import asyncio; from mudd.database import init_database; asyncio.run(init_database())"
