+++ mudd/cogs/look.py
@@ -7,7 +7,7 @@ import asyncpg
 from discord import Interaction, app_commands
 from discord.ext import commands
 
-from mudd.services.entity_matcher import (
+from mudd.matching.entity_matcher import (
     get_focus_aware_autocomplete_entities,
     match_entity_by_prefix,
 )
@@ -16,7 +16,7 @@ from mudd.services.rendering import RenderingService, TemplateRenderError
 if TYPE_CHECKING:
     from mudd.services.entity import EntityService
     from mudd.services.focus_context import FocusContextService
-    from mudd.services.visibility_protocol import VisibilityServiceProtocol
+    from mudd.services.visibility import VisibilityServiceProtocol
 
 logger = logging.getLogger(__name__)
