+++ mudd/loaders/__init__.py
@@ -1,0 +1,1 @@
+"""Loaders for syncing data from source files to the database."""
