+++ mudd/matching/__init__.py
@@ -1,0 +1,1 @@
+"""Matching functions for entities and verbs."""
