+++ mudd/matching/verb_matcher.py
@@ -4,7 +4,7 @@ import logging
 
 import asyncpg
 
-from mudd.services.verb_action import VerbAction
+from mudd.types import VerbAction
 
 logger = logging.getLogger(__name__)
