+++ mudd/services/CLAUDE.md
@@ -1,0 +1,22 @@
+# Services
+
+This package contains **DI service classes** that manage runtime state, caching, and database access.
+
+## What makes a service
+A service is a class that:
+- Receives dependencies (database pool, other services) via constructor injection
+- Manages state or caching
+- Provides methods for data access or business logic
+- Is instantiated once and passed to components that need it
+
+## What belongs here
+- EntityService - entity resolution with caching
+- FocusContextService - user focus state management
+- VisibilityService - user location and Discord visibility
+- RenderingService - template rendering with caching
+
+## What does NOT belong here
+- Module-level functions (use mudd/database.py or dedicated modules)
+- Data loading/sync functions (use loaders/)
+- Pure matching functions (use matching/)
+- Type definitions (use mudd/types.py)
