+++ tests/loaders/test_entity_loader.py
@@ -18,8 +18,8 @@ Tests:
 import pytest
 import pytest_asyncio
 
-from mudd.services.entity_loader import _validate_and_sort_entities, sync_entities
-from mudd.services.zone_loader import (
+from mudd.loaders.entity_loader import _validate_and_sort_entities, sync_entities
+from mudd.loaders.zone_loader import (
     Entity,
     load_entities_from_rec,
     load_rooms_from_rec,
