+++ tests/unit/test_entity_matcher.py
@@ -2,11 +2,11 @@
 
 from uuid import uuid4
 
-from mudd.services.entity import EntityInstance, ResolvedEntity
-from mudd.services.entity_matcher import (
+from mudd.matching.entity_matcher import (
     MatchResult,
     match_entity_by_prefix,
 )
+from mudd.services.entity import EntityInstance, ResolvedEntity
 
 
 def make_entity(
@@ -166,7 +166,7 @@ class TestMatchResult:
 
     def test_is_unique_with_multiple_matches(self):
         """is_unique returns False for multiple matches."""
-        from mudd.services.entity_matcher import EntityMatch
+        from mudd.matching.entity_matcher import EntityMatch
 
         table = make_entity("table", "Table")
         chair = make_entity("chair", "Chair")
@@ -180,7 +180,7 @@ class TestMatchResult:
 
     def test_is_ambiguous_with_multiple_matches(self):
         """is_ambiguous returns True for multiple matches."""
-        from mudd.services.entity_matcher import EntityMatch
+        from mudd.matching.entity_matcher import EntityMatch
 
         table = make_entity("table", "Table")
         chair = make_entity("chair", "Chair")
